requirements.txt
groq
python-dotenv
python-docx
requests
beautifulsoup4
gradio

